fix deblending
Always run the deblender subtask, but set
    deblend_config.propagateAllPeaks = True

so that sub-peaks are effectively found and reported in the source list.

Now the option is whether or not to actually deblend, which involves
using the noise replacer.

The option is now simply called "deblend" in the keywords and
configuration option.

default is current deblend=False