send a noise image through for deblending
For metacal we should generate a noise image so that the exact noise
field is used for all versions of the metacal images.  The assumption is
that, because these noise data should contain no signal, metacal is not
calibrating it.  Thus it doesn't matter whether or not the noise field
is representative of the full covariance of the true image noise.
Rather the motivation for making the field the same for all metacal
images is to reduce variance in the calculation of the response