Use correct catalog, attach footprints
The code runs and produces measurements when there is blending.

This still uses the footprints to store models, which fred has said
should work but is probably slower that getting the models directly.
I'm sticking with this for now because I don't currently know how to
subtract models once I have them.

Still to do is check this is doing something sensible with the model
subtraction, insertion etc.